backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
groq>=0.9.0
libcst>=1.4.0
pydantic>=2.9.0
python-dotenv>=1.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
hypothesis>=6.100.0
anyio>=4.4.0
hindsight-client
